requirements.txt
Flask==1.1.2
Jinja2==2.11.2
numpy==1.18.5
pickleshare==0.7.5
Pillow==7.2.0
requests==2.24.0
requests-oauthlib==1.3.0
https://download.pytorch.org/whl/cpu/torch-1.0.1.post2-cp36-cp36m-linux_x86_64.whl
gunicorn==19.9.0
